fix counting adults and children for order tickets with same ticketType
